+++ src/XYSprite.cpp
@@ -11,19 +11,19 @@
 XYSprite::XYSprite(int x, int y, Sprite::RenderMode mode, const Sprite sprite) :
 mode_(mode), sprite_(sprite), x_(x), y_(y) {}
 
-void XYSprite::render(int page_num, uint8_t *page) const {
+void XYSprite::render(uint8_t page_num, uint8_t *page) const {
 	render(page_num, page, (Sprite::RenderMode) mode_);
 }
 
-void XYSprite::render(int page_num, uint8_t *page, Sprite::RenderMode mode) const {
+void XYSprite::render(uint8_t page_num, uint8_t *page, Sprite::RenderMode mode) const {
 	if (!sprite_.raster())
 		return;
-	int page_y = page_num << 3;
+	uint8_t page_y = page_num << 3;
 	if (y_ + sprite_.height() < page_y || y_ >= page_y + 8)
 		return;
 	if (x_ + sprite_.width() < 0 || x_ >= 128)
 		return;
-	int width = sprite_.width();
+	uint8_t width = sprite_.width();
 	if (x_ + width > 128)
 		width = 128 - x_;
 	sprite_.render(mode, 0, page_y - y_, width, page + x_);
